python_backend/requirements.txt
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
python-dotenv>=1.0.0
psycopg2-binary>=2.9.9
redis>=5.0.0
python-multipart>=0.0.6
openpyxl>=3.1.2
pandas>=2.1.0
aiofiles>=23.2.1
httpx>=0.25.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
litellm>=1.40.0

# Evaluation Pipeline Dependencies
google-generativeai>=0.3.0
python-pptx>=0.6.21
python-docx>=1.0.0
Pillow>=10.0.0
opencv-python>=4.8.0
tqdm>=4.66.0
tenacity>=8.2.0
reportlab>=4.0.0
